Move tracking utils to scenes directory
